Move wallet reducer
